+++ src/ui/routes.tsx
@@ -2,15 +2,18 @@
 import { HashRouter, Routes, Route } from "react-router-dom";
 import Player from "./Player";
 import Home from "./pages/Home";
+import AllTracks from "./pages/AllTracks";
 
 const AppRoutes = () => {
 	console.log(window.location.pathname);
+	console.log(window.location.hash);
 
 	return (
 		<HashRouter>
 			<Routes>
 				<Route path="/" element={<Player />}>
 					<Route index element={<Home />} />
+					<Route path="all-tracks" element={<AllTracks />} />
 				</Route>
 			</Routes>
 		</HashRouter>